styles: refactor and update styles
