+++ src/client/components/common/Alert/Alert.tsx
@@ -5,9 +5,11 @@ import ErrorIcon from './error.svg';
 import WarningIcon from './warning.svg';
 import styles from './alert.module.scss';
 import Image from 'next/image';
+import classNames from 'classnames';
 
 type AlertProps = {
   severity: Severity;
+  className?: string;
 } & PropsWithChildren;
 
 const STYLES_MAP: Record<Severity, keyof typeof styles> = {
@@ -22,9 +24,9 @@ const ICON_MAP: Record<Severity, any> = {
   success: SuccessIcon,
 };
 
-const Alert: FunctionComponent<AlertProps> = ({ severity, children }) => {
+const Alert: FunctionComponent<AlertProps> = ({ severity, children, className }) => {
   return (
-    <div className={`${styles.alert} ${STYLES_MAP[severity]}`}>
+    <div className={classNames(styles.alert, STYLES_MAP[severity], className)}>
       <div className={styles.iconWrapper}>
         <Image src={ICON_MAP[severity]} alt="" />
       </div>
