+++ src/pages/credential-stuffing/credentialStuffing.module.scss
@@ -1,0 +1,9 @@
+.credentialStuffingForm {
+  .alert {
+    margin: rem(24px) 0;
+  }
+  button[type='submit'] {
+    margin: 0 auto;
+    margin-top: rem(24px);
+  }
+}
