+++ src/pages/payment-fraud/index.tsx
@@ -6,7 +6,9 @@ import { USE_CASES } from '../../client/components/common/content';
 import Button from '../../client/components/common/Button';
 
 import styles from './paymentFraud.module.scss';
+import formStyles from '../../styles/forms.module.scss';
 import Alert from '../../client/components/common/Alert/Alert';
+import classNames from 'classnames';
 
 export default function Index() {
   const visitorDataQuery = useVisitorData({
@@ -63,8 +65,8 @@ export default function Index() {
 
   return (
     <UseCaseWrapper useCase={USE_CASES.paymentFraud} contentSx={{ maxWidth: 'none' }}>
-      <div className={styles.wrapper}>
-        <form onSubmit={handleSubmit} className={styles.paymentForm}>
+      <div className={formStyles.wrapper}>
+        <form onSubmit={handleSubmit} className={classNames(formStyles.useCaseForm, styles.paymentForm)}>
           <label>Card Number</label>
           <input
             type="text"
@@ -102,7 +104,7 @@ export default function Index() {
           <hr />
 
           <div className={styles.checkboxes}>
-            <label className={styles.checkboxLabel}>
+            <label className={formStyles.checkboxLabel}>
               <input
                 type="checkbox"
                 name="applyChargeback"
@@ -111,7 +113,7 @@ export default function Index() {
               Ask for chargeback after purchase
             </label>
 
-            <label className={styles.checkboxLabel}>
+            <label className={formStyles.checkboxLabel}>
               <input
                 type="checkbox"
                 name="usingStolenCard"
@@ -124,7 +126,7 @@ export default function Index() {
           </div>
 
           {httpResponseStatus ? <Alert severity={severity}>{orderStatusMessage}</Alert> : null}
-          <Button disabled={isWaitingForResponse} size="large" type="submit" className={styles.submitButton}>
+          <Button disabled={isWaitingForResponse} size="large" type="submit">
             {isWaitingForResponse ? 'Hold on, doing magic...' : 'Place Order'}
           </Button>
         </form>
